+++ packages/tailwind/integrations/nextjs/src/app/page.tsx
@@ -1,0 +1,7 @@
+import { renderAsync } from "@react-email/components";
+import { VercelInviteUserEmail } from "../../emails/vercel-invite-user";
+
+export default async function Home() {
+  const emailHtml = await renderAsync(<VercelInviteUserEmail />);
+  return <div dangerouslySetInnerHTML={{ __html: emailHtml }} />;
+}